created dashboard and moved all the component related, also fixed some bugs
