+++ client/src/_lib/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import { signIn } from "../../auth";
+import { signIn } from "../auth";
 import { AuthError } from "next-auth";
 
 export async function authenticate(
