+++ client/src/app/dashboard/article/edit/[id]/page.tsx
@@ -1,24 +1,30 @@
 "use client";
 import { useState, useContext, useEffect, ChangeEvent } from "react";
-import { GlobalContext } from "../../global-context";
+import { GlobalContext } from "../../../global-context";
+import { SingleArticleResponse } from "../../../ownTypes";
 import {
   Box,
-  Chip,
   FormControl,
   Input,
   InputLabel,
   LinearProgress,
   Typography,
+  Chip,
 } from "@mui/material";
+import Image from "next/image";
 import SaveIcon from "@mui/icons-material/Save";
 import { Button } from "@mui/material";
 import { EditorProps } from "react-draft-wysiwyg";
-import { EditorState, ContentState, convertToRaw } from "draft-js";
-import { toolbarconfig } from "../../../_lib/editor-toolbar-config";
+import {
+  EditorState,
+  ContentState,
+  convertFromHTML,
+  convertToRaw,
+} from "draft-js";
+import { toolbarconfig } from "../../../../../_lib/editor-toolbar-config";
 import draftToHtml from "draftjs-to-html";
 import "react-draft-wysiwyg/dist/react-draft-wysiwyg.css";
 import Modal from "@mui/material/Modal";
-import { useRouter } from "next/navigation";
 import dynamic from "next/dynamic";
 
 const backendBaseURL = process.env.BACKEND_URL;
@@ -40,26 +46,42 @@ const Editor = dynamic<EditorProps>(
   { ssr: false }
 );
 
-export default function Add() {
+export default function Edit({
+  //@ts-ignore
+  params: { id },
+}) {
   const { setGlobalState } = useContext(GlobalContext);
-  const router = useRouter();
-
   useEffect(() => {
     if (setGlobalState) {
-      setGlobalState({ title: "Agregar Artículo" });
+      setGlobalState({ title: "Editar Artículo" });
     }
   }, [setGlobalState]);
 
+  const [dataArticle, setDataArticle] = useState<SingleArticleResponse>();
   const [title, setTitle] = useState<string>();
   const [category, setCategory] = useState<string>();
   const [dataCategories, setDataCategories] = useState<string[]>();
-
   const [editorState, setEditorState] = useState(() => {
-    const content = ContentState.createFromText("");
+    const content = ContentState.createFromText("Content is loading...");
     return EditorState.createWithContent(content);
   });
   const [open, setOpen] = useState(false);
   const [file, setFile] = useState<File>();
+
+  useEffect(() => {
+    const loadArticles = async () => {
+      try {
+        const response = await fetch(`${backendBaseURL}/articles/${id}`, {
+          mode: "cors",
+        });
+        setDataArticle(await response?.json());
+      } catch (err) {
+        console.log(err, `Could not load the article id:${id}`);
+      }
+    };
+    loadArticles();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   useEffect(() => {
     const loadCategories = async () => {
@@ -76,7 +98,24 @@ export default function Add() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  const addArticle = async () => {
+  const onCategoryChange = (
+    event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
+  ) => {
+    setCategory(event.target.value);
+  };
+
+  useEffect(() => {
+    setTitle(dataArticle?.title);
+
+    const blocksFromHTML = convertFromHTML(dataArticle?.content || "");
+    const content = ContentState.createFromBlockArray(
+      blocksFromHTML.contentBlocks,
+      blocksFromHTML.entityMap
+    );
+    setEditorState(EditorState.createWithContent(content));
+  }, [dataArticle]);
+
+  const saveArticle = async () => {
     let bufferImage = null;
     if (file) {
       const bytes = await file.arrayBuffer();
@@ -87,14 +126,15 @@ export default function Add() {
     const data = {
       title: title ? title.trim() : "",
       content: convertContentToHTML(editorState),
-      image: bufferImage,
+      image: file ? bufferImage : dataArticle?.image,
       category: category ? category.trim() : "",
     };
-    fetch(`${backendBaseURL}/articles/`, {
+
+    fetch(`${backendBaseURL}/articles/${id}`, {
       headers: {
         "Content-Type": "application/json", // Set the content type if you're sending JSON data
       },
-      method: "POST",
+      method: "PATCH",
       mode: "cors",
       body: JSON.stringify(data),
     })
@@ -107,10 +147,8 @@ export default function Add() {
 
   const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    if (!file) return;
-
     try {
-      addArticle();
+      saveArticle();
     } catch (e: any) {
       // Handle errors here
       console.error(e);
@@ -118,25 +156,13 @@ export default function Add() {
   };
 
   const handleOpenModal = () => setOpen(true);
-  const handleCloseModal = () => {
-    setOpen(false);
-    setTimeout(() => {
-      router.push("/");
-    }, 500);
-  };
+  const handleCloseModal = () => setOpen(false);
 
   const onTitleChange = (
     event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
   ) => {
     setTitle(event.target.value);
   };
-
-  const onCategoryChange = (
-    event: ChangeEvent<HTMLTextAreaElement | HTMLInputElement>
-  ) => {
-    setCategory(event.target.value);
-  };
-
   const onEditorChange = (newEditorState: any) => {
     setEditorState(newEditorState);
   };
@@ -149,7 +175,7 @@ export default function Add() {
     return markup;
   }
 
-  return dataCategories ? (
+  return dataArticle && dataCategories ? (
     <>
       <form onSubmit={onSubmit}>
         <FormControl fullWidth>
@@ -165,7 +191,6 @@ export default function Add() {
             onChange={(e) => onTitleChange(e)}
           />
         </FormControl>
-
         <FormControl sx={{ marginTop: 5 }}>
           <InputLabel htmlFor="category" focused shrink>
             Categoria
@@ -174,6 +199,7 @@ export default function Add() {
             id="category"
             aria-describedby="Categoría del articulo"
             value={category}
+            defaultValue={dataArticle.category}
             fullWidth
             required
             onChange={(e) => onCategoryChange(e)}
@@ -190,7 +216,6 @@ export default function Add() {
             ))}
           </Box>
         </FormControl>
-
         <FormControl fullWidth sx={{ marginTop: 5, marginBottom: 5 }}>
           <InputLabel htmlFor="content" focused shrink>
             Contenido
@@ -208,6 +233,12 @@ export default function Add() {
         </FormControl>
 
         <FormControl fullWidth sx={{ marginTop: 5, marginBottom: 5 }}>
+          <Image
+            width={80}
+            height={80}
+            src={`data:image/png;base64,${dataArticle.image}`}
+            alt={dataArticle.title}
+          />
           <Input
             id="image"
             aria-describedby="Imagen para el articulo"
@@ -240,7 +271,7 @@ export default function Add() {
             component="h2"
             sx={{ textAlign: "center" }}
           >
-            Artículo agregado satisfactoriamente!
+            Artículo actualizado satisfactoriamente!
           </Typography>
           <Button onClick={handleCloseModal}>Cerrar</Button>
         </Box>
