+++ client/src/app/dashboard/global-context.tsx
