+++ client/src/app/dashboard/header.tsx
@@ -1,14 +1,13 @@
-'use client'
-import React, {useContext} from 'react';
-import AppBar from '@mui/material/AppBar';
-import Grid from '@mui/material/Grid';
-import HelpIcon from '@mui/icons-material/Help';
-import IconButton from '@mui/material/IconButton';
-import Toolbar from '@mui/material/Toolbar';
-import Tooltip from '@mui/material/Tooltip';
-import Typography from '@mui/material/Typography';
-import {GlobalContext} from './global-context';
-
+"use client";
+import React, { useContext } from "react";
+import AppBar from "@mui/material/AppBar";
+import Grid from "@mui/material/Grid";
+import HelpIcon from "@mui/icons-material/Help";
+import IconButton from "@mui/material/IconButton";
+import Toolbar from "@mui/material/Toolbar";
+import Tooltip from "@mui/material/Tooltip";
+import Typography from "@mui/material/Typography";
+import { GlobalContext } from "./global-context";
 
 export default function Header() {
   const globalContext = useContext(GlobalContext);
@@ -22,7 +21,14 @@ export default function Header() {
         sx={{ zIndex: 0 }}
       >
         <Toolbar>
-          <Grid container alignItems="center" justifyContent="center" alignContent="center" height={130} spacing={1}>
+          <Grid
+            container
+            alignItems="center"
+            justifyContent="center"
+            alignContent="center"
+            height={130}
+            spacing={1}
+          >
             <Grid item lg>
               <Typography color="inherit" variant="h5" component="h1">
                 {globalContext?.globalState?.title}
