+++ client/src/app/dashboard/template.tsx
@@ -1,23 +1,23 @@
-'use client'
+"use client";
 //https://github.com/mui/material-ui/blob/master/docs/src/pages/premium-themes/
-import * as React from 'react';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import useMediaQuery from '@mui/material/useMediaQuery';
-import CssBaseline from '@mui/material/CssBaseline';
-import Box from '@mui/material/Box';
-import Typography from '@mui/material/Typography';
-import Link from '@mui/material/Link';
-import Navigator from './navigator';
-import Header from './header';
-import GlobalProvider from './global-context'
+import * as React from "react";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import useMediaQuery from "@mui/material/useMediaQuery";
+import CssBaseline from "@mui/material/CssBaseline";
+import Box from "@mui/material/Box";
+import Typography from "@mui/material/Typography";
+import Link from "@mui/material/Link";
+import Navigator from "./navigator";
+import Header from "./header";
+import GlobalProvider from "./global-context";
 
 function Copyright() {
   return (
     <Typography variant="body2" color="text.secondary" align="center">
-      {'Copyright © '}
+      {"Copyright © "}
       <Link color="inherit" href="https://samilabud.netlify.app/#about">
-      Samil Abud
-      </Link>{' '}
+        Samil Abud
+      </Link>{" "}
       {new Date().getFullYear()}.
     </Typography>
   );
@@ -26,9 +26,9 @@ function Copyright() {
 let theme = createTheme({
   palette: {
     primary: {
-      light: '#63ccff',
-      main: '#009be5',
-      dark: '#006db3',
+      light: "#63ccff",
+      main: "#009be5",
+      dark: "#006db3",
     },
   },
   typography: {
@@ -61,19 +61,19 @@ theme = {
     MuiDrawer: {
       styleOverrides: {
         paper: {
-          backgroundColor: '#081627',
+          backgroundColor: "#081627",
         },
       },
     },
     MuiButton: {
       styleOverrides: {
         root: {
-          textTransform: 'none',
+          textTransform: "none",
         },
         contained: {
-          boxShadow: 'none',
-          '&:active': {
-            boxShadow: 'none',
+          boxShadow: "none",
+          "&:active": {
+            boxShadow: "none",
           },
         },
       },
@@ -94,11 +94,11 @@ theme = {
     MuiTab: {
       styleOverrides: {
         root: {
-          textTransform: 'none',
-          margin: '0 16px',
+          textTransform: "none",
+          margin: "0 16px",
           minWidth: 0,
           padding: 0,
-          [theme.breakpoints.up('md')]: {
+          [theme.breakpoints.up("md")]: {
             padding: 0,
             minWidth: 0,
           },
@@ -122,15 +122,15 @@ theme = {
     MuiDivider: {
       styleOverrides: {
         root: {
-          backgroundColor: 'rgb(255,255,255,0.15)',
+          backgroundColor: "rgb(255,255,255,0.15)",
         },
       },
     },
     MuiListItemButton: {
       styleOverrides: {
         root: {
-          '&.Mui-selected': {
-            color: '#4fc3f7',
+          "&.Mui-selected": {
+            color: "#4fc3f7",
           },
         },
       },
@@ -146,10 +146,10 @@ theme = {
     MuiListItemIcon: {
       styleOverrides: {
         root: {
-          color: 'inherit',
-          minWidth: 'auto',
+          color: "inherit",
+          minWidth: "auto",
           marginRight: theme.spacing(2),
-          '& svg': {
+          "& svg": {
             fontSize: 20,
           },
         },
@@ -170,7 +170,7 @@ const drawerWidth = 256;
 
 export default function Template({ children }: { children: React.ReactNode }) {
   const [mobileOpen, setMobileOpen] = React.useState(false);
-  const isSmUp = useMediaQuery(theme.breakpoints.up('sm'));
+  const isSmUp = useMediaQuery(theme.breakpoints.up("sm"));
 
   const handleDrawerToggle = () => {
     setMobileOpen(!mobileOpen);
@@ -178,37 +178,40 @@ export default function Template({ children }: { children: React.ReactNode }) {
 
   return (
     <ThemeProvider theme={theme}>
-        <GlobalProvider>
-        <Box sx={{ display: 'flex', minHeight: '100vh' }}>
-            <CssBaseline />
-            <Box
+      <GlobalProvider>
+        <Box sx={{ display: "flex", minHeight: "100vh" }}>
+          <CssBaseline />
+          <Box
             component="nav"
             sx={{ width: { sm: drawerWidth }, flexShrink: { sm: 0 } }}
-            >
+          >
             {isSmUp ? null : (
-                <Navigator
+              <Navigator
                 PaperProps={{ style: { width: drawerWidth } }}
                 variant="temporary"
                 open={mobileOpen}
                 onClose={handleDrawerToggle}
-                />
+              />
             )}
             <Navigator
-                PaperProps={{ style: { width: drawerWidth } }}
-                sx={{ display: { sm: 'block', xs: 'none' } }}
+              PaperProps={{ style: { width: drawerWidth } }}
+              sx={{ display: { sm: "block", xs: "none" } }}
             />
-            </Box>
-            <Box sx={{ flex: 1, display: 'flex', flexDirection: 'column' }}>
+          </Box>
+          <Box sx={{ flex: 1, display: "flex", flexDirection: "column" }}>
             <Header />
-            <Box component="main" sx={{ flex: 1, py: 6, px: 4, bgcolor: '#eaeff1' }}>
-                {children}
+            <Box
+              component="main"
+              sx={{ flex: 1, py: 6, px: 4, bgcolor: "#eaeff1" }}
+            >
+              {children}
             </Box>
-            <Box component="footer" sx={{ p: 2, bgcolor: '#eaeff1' }}>
-                <Copyright />
+            <Box component="footer" sx={{ p: 2, bgcolor: "#eaeff1" }}>
+              <Copyright />
             </Box>
-            </Box>
+          </Box>
         </Box>
-    </GlobalProvider>
-   </ThemeProvider>
+      </GlobalProvider>
+    </ThemeProvider>
   );
 }
