+++ client/src/app/globals.css
@@ -15,13 +15,3 @@
     --background-end-rgb: 0, 0, 0;
   }
 }
-/* 
-body {
-  color: rgb(var(--foreground-rgb));
-  background: linear-gradient(
-      to bottom,
-      transparent,
-      rgb(var(--background-end-rgb))
-    )
-    rgb(var(--background-start-rgb));
-} */
