+++ client/src/auth.ts
@@ -21,18 +21,27 @@ export const { auth, signIn, signOut } = NextAuth({
   providers: [
     Credentials({
       async authorize(credentials) {
-        console.log({ z: z.string() });
         const parsedCredentials = z
           .object({ email: z.string().email(), password: z.string().min(6) })
           .safeParse(credentials);
 
         if (parsedCredentials.success) {
           const { email, password } = parsedCredentials.data;
-          const user = await getUser(email);
-          if (!user) return null;
-          const passwordsMatch = await bcrypt.compare(password, user.password);
+          // const user = await getUser(email);
+          // if (!user) return null;
+          // const passwordsMatch = await bcrypt.compare(password, user.password);
 
-          if (passwordsMatch) return user;
+          // if (passwordsMatch) return user;
+          const userMock = {
+            id: "1",
+            name: "Paola Guzmán",
+            email: "admin@lactaconsejos.com",
+            password: "123456",
+          };
+
+          if (email === userMock.email && password === userMock.password) {
+            return userMock;
+          }
         }
 
         console.log("Invalid credentials");
